Updated holdout to support experimental setup for synthetic lethality

+++ bindings/python/src/metrics.rs
@@ -175,7 +175,7 @@ impl EnsmallenGraph {
 
     #[text_signature = "(self)"]
     /// Returns the mean number of types for each edge.
-    fn get_mean_number_of_types_for_edge(&self) -> usize {
+    fn get_mean_number_of_types_for_edge(&self) -> f64 {
         self.graph.get_mean_number_of_types_for_edge()
     }
 }
